requirements.txt
flask==3.1.*
customtkinter==5.*
qrcode[pil]==8.*
openpyxl==3.*
Pillow==11.*
